refactor: modularize Kowalski agent skills into dedicated analysis and communication modules
